third_party/blink/web_tests/wpt_internal/geolocation-api/reentrant-success.https.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/testharness-adapter.js"></script>
</head>
<body>
<script type="module">
import {GeolocationMock} from './resources/geolocation-mock.js';

description("Tests that reentrant calls to Geolocation methods from the success callback are OK.");

let mockLatitude = 51.478;
let mockLongitude = -0.166;
let mockAccuracy = 100.0;

let position;

const mock = new GeolocationMock();
mock.setGeolocationPermission(true);
mock.setGeolocationPosition(mockLatitude,
                            mockLongitude,
                            mockAccuracy);

let successCallbackInvoked = false;
navigator.geolocation.getCurrentPosition(function(p) {
    if (successCallbackInvoked) {
        testFailed('Success callback invoked unexpectedly');
        finishJSTest();
    }
    successCallbackInvoked = true;

    position = p;
    assert_equals(position.coords.latitude, mockLatitude);
    assert_equals(position.coords.longitude, mockLongitude);
    assert_equals(position.coords.accuracy, mockAccuracy);
    continueTest();
}, function(e) {
    testFailed('Error callback invoked unexpectedly');
    finishJSTest();
});

function continueTest() {
    mock.setGeolocationPosition(++mockLatitude,
                                ++mockLongitude,
                                ++mockAccuracy);

    navigator.geolocation.getCurrentPosition(function(p) {
        position = p;
        assert_equals(position.coords.latitude, mockLatitude);
        assert_equals(position.coords.longitude, mockLongitude);
        assert_equals(position.coords.accuracy, mockAccuracy);
        finishJSTest();
    }, function(e) {
        testFailed('Error callback invoked unexpectedly');
        finishJSTest();
    });
}
</script>
</body>
</html>
